backend/requirements.txt
Django>=4.2
djangorestframework
pandas
reportlab
djangorestframework-simplejwt

PyQt5
requests
matplotlib
